feat : 현재위치좌표 장소검색 로직 수정 및 검색결과에 따른 조건부렌더링 추가
